refactor(Modules): Move module to it's own folder
