+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './api/Webpack.js';
+export * from './api/modules/Webpack.js';
